switch to per frame processing
