+++ src/dsp/instrument_processor.cpp
@@ -69,15 +69,6 @@ InstrumentProcessor::InstrumentProcessor(std::span<const BarelyControlOverride> 
   params_.rng = &rng;
 }
 
-void InstrumentProcessor::Process(float* BARELY_RESTRICT delay_samples,
-                                  float* BARELY_RESTRICT output_samples, int output_channel_count,
-                                  int output_frame_count) noexcept {
-  for (VoiceState& voice_state : voice_states_) {
-    voice_callback_(voice_state.voice, params_, delay_samples, output_samples, output_channel_count,
-                    output_frame_count);
-  }
-}
-
 // NOLINTNEXTLINE(bugprone-exception-escape)
 void InstrumentProcessor::SetControl(ControlType type, float value) noexcept {
   switch (type) {
